refactor HoH and dependents to use the same table as apply to household component
